requirements.txt
Flask
psycopg
psycopg[binary]